# modules/reservation/reservation-headless/reservation-headless-impl/rest-openapi.yaml
components:
  schemas:
    Time:
      description: Contains time data.
      properties:
        hours:
          type: string
        minutes:
          type: string
        ampm:
          type: string
      required: 
        - hours
        - minutes
        - ampm
      type: object
...
    Booking:
      description: Contains information about the booking.
      properties:
        bookingId:
          description: The primary key for the Booking entity.
          type: integer
          format: int64
          readOnly: true
        title:
          description: Title of the booking event.
          type: string
        date:
          description: Date of the booking event.
          type: string
          pattern: ^\d{4}\-(0?[1-9]|1[012])\-(0?[1-9]|[12][0-9]|3[01])$
        startTime:
          description: Start time of the booking event.
          $ref: "#/components/schemas/Time"
        endTime:
          description: End time of the booking event.
          $ref: "#/components/schemas/Time"
        roomId:
          description: Id associated with the room.
          type: integer
          format: int64
        roomName:
          description: Name associated with the room.
          type: string
          readOnly: true
        officeId:
          description: Id associated with the office.
          type: integer
          format: int64
          readOnly: true
        officeName:
          description: Name associated with the room.
          type: string
          readOnly: true
        description:
          description: A description of the booking event.
          type: string
        userId:
          description: The user who created the booking.
          type: integer
          format: int64
          readOnly: true
        userName:
          description: Name of associated user.
          type: string 
          readOnly: true
        participantIds:
          description: Primary keys of the participants for this booking.
          type: array
          items:
            type: integer
            format: int64
      required: 
        - title
        - date
        - startTime
        - endTime
        - roomId
        - description
      type: object
openapi: 3.0.1
info:
  description: "API for accessing Reservation details."
  license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  title: "Headless Reservation"
  version: v1.0
paths:
...
  "/bookings":
    post:
      operationId: postBooking
      tags: ["Booking"]
      description: Create a new booking.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Booking"
          application/xml:
            schema:
              $ref: "#/components/schemas/Booking"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
            application/xml:
              schema:
                $ref: "#/components/schemas/Booking"
    get:
      operationId: getBookingsPage
      tags: ["Booking"]
      description: Retrieve the list of bookings. Results can be paginated, filtered, searched and sorted.
      parameters:
        - in: query
          name: filter
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Booking"
                type: array
            application/xml:
              schema:
                items:
                  $ref: "#/components/schemas/Booking"
                type: array
  "/bookings/{bookingId}":
    get:
      operationId: getBooking
      tags: ["Booking"]
      description: Retrieves the booking via its ID.
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
            application/xml:
              schema:
                $ref: "#/components/schemas/Booking"
    put: 
      operationId: putBooking
      tags: ["Booking"]
      description: Updates and overwrites the booking.
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Booking"
          application/xml:
            schema:
              $ref: "#/components/schemas/Booking"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
            application/xml:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: ["Booking"]
      description: Deletes the booking and returns a 204 if the operation succeeds.
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: ""
          content:
            application/json: {}
            application/xml: {}